+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 0.1.5.dev
 ---------
 
+- New convenience methods :func:`Grid.record_positions`
+  and :func:`Space.record_positions` have been added to record
+  agent positions.
 - :func:`Model.run` can now continue simulations that have already been run.
   The steps defined in the argument 'steps' now reflect additional steps,
   which will be added to the models current time-step.